Don’t put Sarah on a line by herself
Your use of 33% width image divs with four admins puts Sarah below the
other admins on a line by herself. This was probably unintentional, but
it sends a message that we hold men in higher regard. We don’t.

When more admins are added to the page, 33% styling may be appropriate,
but consider alphabetizing the admins by last name so we wouldn’t have,
say, Sarah and Julie on line 2, and all the men on line 1.

Thanks!